Changes from code review
* Remove code that has moved out of actions/__init__.py
* Remove unused imports
* Change copyright year of actions/__init__.py since code here is 100% new in this commit.
* Update comments in resolve_action_order
* Make report functions more organized
  * Rename actions.format_report_message() to format_action_status_message() to make it clear it is
    related to Action rather than only to generating the report.
  * Move format_report_section_header() to report.py
